pyproject.toml
#:schema https://json.schemastore.org/pyproject.json
[project]
name = "advent-of-visualizations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "matplotlib>=3.10.7",
  "numpy>=2.3.5",
  "pillow>=12.0.0",
  "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
  "click<=8.2.1",
  "mkdocs-material>=9.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
typeCheckingMode = "standard"
ignore = ["__init__.py"]

[tool.ruff.lint]
ignore = ["ALL"]
